refactor(mail): merge mail storage and search into suite.mail.store
Fold suite.mail.storage and suite.mail.search into a single
suite.mail.store package, mirroring the core suite.store merge: the
Entity/namespace/factory helpers, the whitelisted destroys, and the
email-address index build/rebuild functions now live together, with the
index subpackage moved under store/indexes.

Repoint every importer (patches, api, jmap, utils, contact_card,
jmap_account, mail_message) and the mail_settings.js whitelisted method
paths at the merged package. rebuild_email_address_index no longer needs
a lazy Entity/get_data_store import now that they share a module.
