+++ suite/mail/doctype/mail_message/mail_message.py
@@ -30,8 +30,7 @@ from suite.mail.doctype.user_account.user_account import get_user_for_jmap_accou
 from suite.mail.jmap import get_email_service, get_jmap_connection, get_thread_service
 from suite.mail.jmap.services.mail.email import EmailService
 from suite.mail.jmap.services.mail.mailbox import MailboxService
-from suite.mail.search import get_email_address_index
-from suite.mail.storage import Entity, get_blob_store, get_data_store
+from suite.mail.store import Entity, get_blob_store, get_data_store, get_email_address_index
 from suite.mail.utils import (
 	get_config,
 	log_mail_error,
