--- suite/mail/search/indexes/__init__.py
@@ -1,3 +1,0 @@
-from suite.mail.search.indexes.email_address import EmailAddressIndex
-
-__all__ = ["EmailAddressIndex"]
